Moves inter service calls to their own controller adds inter service call handler for generating tokens
